1.（核心搜尋演算法重構）： 重構了歷史表（History Heuristics）架構，為不同類型的歷史表實作了分級上限（Tiered Ceilings），並將更新步長（Bonus/Malus）改為平滑且不易過早飽和的線性縮放。 2.（進階啟發式與修正項）： 引入了具備多 Ply 偏移的延續歷史修正更新（Continuation Correction History），在非 PV 節點中套用了節點寬度補償（Node Width Scaling），並對齊了相關搜尋參數。 3.（工具鏈與同步）： 優化了對局測試工具（Match Runner）與搜尋統計數據解析器，並確保所有修改已同步至 classical 與 classical_old 模組中。

+++ chess_engine/classical/search_heuristics.py
@@ -84,7 +84,7 @@ def update_continuation_history(context, ply_offset, prev_move, prev_piece, curr
     current_val = context.continuation_history[ply_offset, prev_piece, prev_to, curr_piece, curr_to]
     clamped_bonus = min(max(bonus, -HISTORY_MAX_CONTINUATION), HISTORY_MAX_CONTINUATION)
     
-    new_val = current_val + clamped_bonus - (current_val * abs(clamped_bonus)) // HISTORY_MAX_CONTINUATION
+    new_val = np.int32(current_val) + clamped_bonus - (np.int32(current_val) * abs(clamped_bonus)) // HISTORY_MAX_CONTINUATION
     context.continuation_history[ply_offset, prev_piece, prev_to, curr_piece, curr_to] = new_val
 
 @numba.njit(cache=True, boundscheck=False, fastmath=True)
